chore!: outsource peer dependency handling
BREAKING CHANGE: `mongoose` and `typescript-optional` should not be included in `monguito`'s bundle; clients should explicitly specify themselves on their `package.json` project files. This way, clients can pick the version of `mongoose` that satisfies their purposes, which `monguito` simply reuses. This change is therefore an improvement for clients but also a source of conflict with existing code.